Add files via upload
I completed this Assessment two different ways, this way uses more of the elements from the modules such as: padding, width, position, and CSS specificity. The previous one uses a more advanced tutorial from google where transform is used to center the middle paragraphs on the page no matter the width of the viewing screen.